Switched to Double for CLI argument processing

+++ SourceryFramework/Sources/Parsing/Utils/AnnotationsParser.swift
@@ -452,7 +452,7 @@ public struct AnnotationsParser {
                     return
                 }
 
-                if let number = Float(value) {
+                if let number = Double(value) {
                     append(key: name, value: NSNumber(value: number), to: &annotations)
                 } else {
                     if (value.hasPrefix("'") && value.hasSuffix("'")) || (value.hasPrefix("\"") && value.hasSuffix("\"")) {
